<refact> : ModifyProfile API 분리하기
- issue #195
