+++ src/api/userAPI.js
@@ -1,2 +1,16 @@
-// import BASE_URL from '../utills/baseUrl';
-// const userAPI = {};
+import BASE_URL from '../utils/baseUrl';
+
+const userAPI = {
+  async getMyinfo(token) {
+    const response = await fetch(`${BASE_URL}/user/myinfo`, {
+      method: 'GET',
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    const data = await response.json();
+    return data;
+  },
+};
+
+export default userAPI;
